updated code for LoR status windows
tried a new way of using the method to call the screen to print up. its
now in the characters file instead of the main game loop, so its only
called when the character is there, and it wont show up... it doesnt
like my global stat_font for somereason. it doesnt see it, do i need to
import into both files for it to work?